+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Removed export functionality.
 - Removed blue info button of displayed maps in map settings.
 - Update Chronostratigraphy Codelist entries.
+- Make Lithology a sub-menu of Stratigraphy.
 
 ### Fixed